refactor: Provide the time function separate to the syntax
