+++ cats-effect/src/main/scala/com/avast/sst/catseffect/syntax/TimeSyntax.scala
@@ -1,11 +1,7 @@
 package com.avast.sst.catseffect.syntax
 
-import java.util.concurrent.TimeUnit
-
-import cats.effect.syntax.bracket._
 import cats.effect.{Bracket, Clock}
-import cats.syntax.flatMap._
-import cats.syntax.functor._
+import com.avast.sst.catseffect.TimeUtils
 import com.avast.sst.catseffect.syntax.TimeSyntax.FOps
 
 import scala.concurrent.duration.Duration
@@ -22,17 +18,7 @@ object TimeSyntax {
   final class FOps[F[_], A](private val f: F[A]) extends AnyVal {
 
     /** Measures the time it takes the effect to finish and records it using the provided function. */
-    def time(record: Duration => F[Unit])(implicit F: Bracket[F, Throwable], C: Clock[F]): F[A] = {
-      val unit = TimeUnit.NANOSECONDS
-      for {
-        start <- C.monotonic(unit)
-        result <- f.guarantee {
-          C.monotonic(unit).map(computeTime(start)).flatMap(record)
-        }
-      } yield result
-    }
-
-    private def computeTime(start: Long)(end: Long) = Duration.fromNanos(end - start)
+    def time(record: Duration => F[Unit])(implicit F: Bracket[F, Throwable], C: Clock[F]): F[A] = TimeUtils.time(f)(record)
 
   }
